# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuTramiteFacil API",
        "description": "API REST para gestión de solicitantes y solicitudes de ayudas sociales (DDD + Clean Architecture).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8010",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "POST /api/auth/register\nCrea un usuario y devuelve su token JWT.",
                "operationId": "bcf74df7bb7d038225880aee509dfebc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Macero"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123",
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado; devuelve access_token y datos del usuario"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "description": "POST /api/auth/login\nVerifica credenciales y devuelve un token JWT.",
                "operationId": "ce6088c43b35a3008d6daa5475812c9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "demo@tutramitefacil.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "demo1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso; devuelve access_token"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión (invalida el token JWT)",
                "description": "POST /api/auth/logout\nInvalida el token actual.",
                "operationId": "176013de7f989f3658907372c4d7c8ea",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "description": "GET /api/auth/me\nDevuelve el usuario autenticado (útil para el frontend).",
                "operationId": "cf09cdc2ad8af937c13d88187eed2c62",
                "responses": {
                    "200": {
                        "description": "Datos del usuario autenticado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}